Add color package, move TEColorParameter
